LayoutTests/fast/js/char-at-expected.txt

This is a test of the charAt and charCodeAt string functions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "".charAt() is ""
PASS "".charCodeAt() is NaN.
PASS "".charAt(undefined) is ""
PASS "".charCodeAt(undefined) is NaN.
PASS "".charAt(0) is ""
PASS "".charCodeAt(0) is NaN.
PASS "".charAt(null) is ""
PASS "".charCodeAt(null) is NaN.
PASS "".charAt(false) is ""
PASS "".charCodeAt(false) is NaN.
PASS "".charAt(true) is ""
PASS "".charCodeAt(true) is NaN.
PASS "".charAt(0) is ""
PASS "".charCodeAt(0) is NaN.
PASS "".charAt(0.1) is ""
PASS "".charCodeAt(0.1) is NaN.
PASS "".charAt(999) is ""
PASS "".charCodeAt(999) is NaN.
PASS "".charAt(Infinity) is ""
PASS "".charCodeAt(Infinity) is NaN.
PASS "".charAt(-1) is ""
PASS "".charCodeAt(-1) is NaN.
PASS "".charAt(-Infinity) is ""
PASS "".charCodeAt(-Infinity) is NaN.
PASS "".charAt(NaN) is ""
PASS "".charCodeAt(NaN) is NaN.
PASS "x".charAt() is "x"
PASS "x".charCodeAt() is 120
PASS "x".charAt(undefined) is "x"
PASS "x".charCodeAt(undefined) is 120
PASS "x".charAt(0) is "x"
PASS "x".charCodeAt(0) is 120
PASS "x".charAt(null) is "x"
PASS "x".charCodeAt(null) is 120
PASS "x".charAt(false) is "x"
PASS "x".charCodeAt(false) is 120
PASS "x".charAt(true) is ""
PASS "x".charCodeAt(true) is NaN.
PASS "x".charAt(0) is "x"
PASS "x".charCodeAt(0) is 120
PASS "x".charAt(0.1) is "x"
PASS "x".charCodeAt(0.1) is 120
PASS "x".charAt(999) is ""
PASS "x".charCodeAt(999) is NaN.
PASS "x".charAt(Infinity) is ""
PASS "x".charCodeAt(Infinity) is NaN.
PASS "x".charAt(-1) is ""
PASS "x".charCodeAt(-1) is NaN.
PASS "x".charAt(-Infinity) is ""
PASS "x".charCodeAt(-Infinity) is NaN.
PASS "x".charAt(NaN) is ""
PASS "x".charCodeAt(NaN) is NaN.
PASS "xy".charAt() is "x"
PASS "xy".charCodeAt() is 120
PASS "xy".charAt(undefined) is "x"
PASS "xy".charCodeAt(undefined) is 120
PASS "xy".charAt(0) is "x"
PASS "xy".charCodeAt(0) is 120
PASS "xy".charAt(null) is "x"
PASS "xy".charCodeAt(null) is 120
PASS "xy".charAt(false) is "x"
PASS "xy".charCodeAt(false) is 120
PASS "xy".charAt(true) is "y"
PASS "xy".charCodeAt(true) is 121
PASS "xy".charAt(0) is "x"
PASS "xy".charCodeAt(0) is 120
PASS "xy".charAt(0.1) is "x"
PASS "xy".charCodeAt(0.1) is 120
PASS "xy".charAt(999) is ""
PASS "xy".charCodeAt(999) is NaN.
PASS "xy".charAt(Infinity) is ""
PASS "xy".charCodeAt(Infinity) is NaN.
PASS "xy".charAt(-1) is ""
PASS "xy".charCodeAt(-1) is NaN.
PASS "xy".charAt(-Infinity) is ""
PASS "xy".charCodeAt(-Infinity) is NaN.
PASS "xy".charAt(NaN) is ""
PASS "xy".charCodeAt(NaN) is NaN.
PASS successfullyParsed is true

TEST COMPLETE
